docs: merged demo to landing page
